break out code for new component called TaskList
