--- src/components/TaskItem.tsx
@@ -1,3 +1,0 @@
-import { useState } from "react";
-import { useTaskStore } from "../stores/useTaskStore";
-
